+++ microservicio/dominio/src/main/java/com/ceiba/moto/modelo/dto/DtoMoto.java
@@ -11,7 +11,7 @@ public class DtoMoto {
     private int cc;
     private String marca;
     private String estado;
-    private int descuento;
+    private double descuento;
     private String nombreImagen;
     private String nombreMoto;
     private int cantidad;
